// types/types.d.ts
type Provinces = 'Alberta' | 'British Columbia' | 'Manitoba' | 'New Brunswick' | 'Newfoundland and Labrador' | 'Northwest Territories' | 'Nova Scotia' | 'Nunavut' | 'Ontario' | 'Prince Edward Island' | 'Quebec' | 'Saskatchewan' | 'Yukon';

export type CurrencyCode = typeof currencies[number]['code'];